requirements.txt
asgiref==3.7.2
attrs==23.1.0
awsebcli==3.20.7
bcrypt==4.0.1
blessed==1.20.0
boto3==1.26.144
botocore==1.29.144
cached-property==1.5.2
cement==2.8.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
Django==4.2.1
django-cors-headers==4.0.0
django-environ==0.10.0
django-filter==23.2
django-storages==1.13.2
djangorestframework==3.14.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==3.2.0
Markdown==3.4.3
MarkupSafe==2.1.2
paramiko==3.2.0
pathspec==0.10.1
pdf2image==1.16.3
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
PyNaCl==1.5.0
pypdf==3.9.1
PyPDF4==1.27.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML>=5.3.1,<=6.0
requests==2.31.0
s3transfer==0.6.1
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.4
termcolor==1.1.0
texttable==1.6.7
uritemplate==4.1.1
urllib3==1.26.16
wcwidth==0.1.9
websocket-client==0.59.0
